[#2] refactor: src renamed, src->app
