+++ tests/test_utils/test_cf_request_handler.py
@@ -1,7 +1,7 @@
 import pytest
 
-from src.constant import Constant
-from src.utils.cf_request_handler import CFRequestHandler
+from app.constant import Constant
+from app.utils.cf_request_handler import CFRequestHandler
 
 
 class TestCFReqHandler:
